+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [Notification](src/common/Notification)
 - [Fields](src/fields)
   - [InputField](src/fields/InputField)
+  - [CheckboxField](src/fields/CheckboxField)
 - [hooks](src/hooks)
   - [useForm](src/hooks/useForm.ts)
   - [useFormValidation](src/hooks/useFormValidation.ts)